backend/requirements.txt
# --- Web Framework & Server ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# --- Database & Migrations ---
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# --- Authentication & Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# --- Validation & Settings ---
pydantic-settings==2.1.0
email-validator==2.1.0

# --- Testing & Development ---
pytest==8.0.0
httpx==0.26.0
black==24.1.1
isort==5.13.2